Time: N/A (0%), Space: N/A (0%) - LeetHub

+++ 0018-4sum/0018-4sum.cpp
@@ -3,13 +3,13 @@ public:
     vector<vector<int>> fourSum(vector<int>& nums, int target) {
         vector<vector<int>>ans;
         sort(nums.begin(),nums.end());
-        for(int i=0;i<nums.size();i++){
+        for(long long int i=0;i<nums.size();i++){
             if(i!=0 && nums[i]==nums[i-1]) continue;
-            for(int j=i+1;j<nums.size()-1;j++){
+            for(long long int j=i+1;j<nums.size()-1;j++){
                  if(j!=i+1 && nums[j]==nums[j-1]) continue;
-                 int k=j+1,l=nums.size()-1;
+                 long long int k=j+1,l=nums.size()-1;
                  while(k<l){
-                     long long sum=nums[i]+nums[j]+nums[k]+nums[l];
+                     long long int sum=nums[i]+nums[j]+nums[k]+nums[l];
                      if(sum<target) k++;
                      else if(sum>target) l--;
                      else{
